# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mint Bucks promotional credit system for Mint Printworks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /credits:
    get:
      operationId: listCredits
      tags: [credits]
      summary: List credits
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: customerId
          in: query
          schema: { type: string }
        - name: expiringDays
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credit"
    post:
      operationId: issueCredit
      tags: [credits]
      summary: Issue Mint Bucks to a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditInput"
      responses:
        "201":
          description: Issued credit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"
  /credits/{id}:
    get:
      operationId: getCredit
      tags: [credits]
      summary: Get credit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Credit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"
        "404":
          description: Not found
    patch:
      operationId: updateCredit
      tags: [credits]
      summary: Update credit (status, expiry, note)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditUpdate"
      responses:
        "200":
          description: Updated credit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"
        "404":
          description: Not found
    delete:
      operationId: deleteCredit
      tags: [credits]
      summary: Cancel / delete a credit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /credits/{id}/redeem:
    post:
      operationId: redeemCredit
      tags: [credits]
      summary: Record a full or partial redemption
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedemptionInput"
      responses:
        "200":
          description: Redemption result with updated credit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionResult"
        "400":
          description: Invalid redemption (e.g. insufficient balance)
        "404":
          description: Credit not found
  /credits/{id}/remind:
    post:
      operationId: sendCreditReminder
      tags: [credits]
      summary: Send email reminder to customer about available credit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reminder sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
        "404":
          description: Credit not found
  /credits/check/{code}:
    get:
      operationId: checkCreditByCode
      tags: [credits]
      summary: Public balance check by credit code (no auth required)
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Credit balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCheckResult"
        "404":
          description: Credit not found
  /credits/{id}/qr:
    get:
      operationId: getCreditQr
      tags: [credits]
      summary: Get QR code PNG for a credit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: QR code PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
  /credits/{id}/certificate:
    get:
      operationId: getCreditCertificate
      tags: [credits]
      summary: Download branded PDF certificate for a credit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PDF certificate
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ── Redemptions ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Credit:
      type: object
      required: [id, customerId, customerName, code, amount, amountRemaining, status, issuedAt, createdAt, updatedAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerCompany:
          type: ["string", "null"]
        code:
          type: string
        amount:
          type: number
        amountRemaining:
          type: number
        status:
          type: string
          enum: [active, partially_redeemed, redeemed, expired, cancelled]
        note:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        issuedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    CreditInput:
      type: object
      required: [customerId, amount]
      properties:
        customerId:
          type: integer
        amount:
          type: number
          minimum: 0.01
        note:
          type: string
        expiresAt:
          type: string
    CreditUpdate:
      type: object
      properties:
        note:
          type: string
        expiresAt:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
    RedemptionInput:
      type: object
      required: [amountApplied]
      properties:
        amountApplied:
          type: number
          minimum: 0.01
        invoiceRef:
          type: string
        note:
          type: string
    Redemption:
      type: object
      required: [id, creditId, customerId, customerName, amountApplied, redeemedAt, createdAt]
      properties:
        id:
          type: integer
        creditId:
          type: integer
        creditCode:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        customerCompany:
          type: ["string", "null"]
        amountApplied:
          type: number
        invoiceRef:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        redeemedAt:
          type: string
        createdAt:
          type: string
    RedemptionResult:
      type: object
      required: [redemption, credit]
      properties:
        redemption:
          $ref: "#/components/schemas/Redemption"
        credit:
          $ref: "#/components/schemas/Credit"
    MessageResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    CreditCheckResult:
      type: object
      required: [code, status, amount, amountRemaining, issuedAt]
      properties:
        code:
          type: string
        status:
          type: string
          enum: [active, partially_redeemed, redeemed, expired, cancelled]
        amount:
          type: number
        amountRemaining:
          type: number
        issuedAt:
          type: string
        expiresAt:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
...