# lib/api-zod/src/generated/types/credit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mint Bucks promotional credit system for Mint Printworks
 * OpenAPI spec version: 0.1.0
 */
import type { CreditStatus } from './creditStatus';

export interface Credit {
  id: number;
  customerId: number;
  customerName: string;
  customerEmail?: string;
  /** @nullable */
  customerCompany?: string | null;
  code: string;
  amount: number;
  amountRemaining: number;
  status: CreditStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  issuedAt: string;
  createdAt: string;
  updatedAt: string;
}