# lib/api-zod/src/generated/types/redemption.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mint Bucks promotional credit system for Mint Printworks
 * OpenAPI spec version: 0.1.0
 */

export interface Redemption {
  id: number;
  creditId: number;
  creditCode?: string;
  customerId: number;
  customerName: string;
  /** @nullable */
  customerCompany?: string | null;
  amountApplied: number;
  /** @nullable */
  invoiceRef?: string | null;
  /** @nullable */
  note?: string | null;
  redeemedAt: string;
  createdAt: string;
}